Added example mnist, augmented cifar10 scripts and densenet implementation.
Deleted old train_cifar10.py script, training for mnist and cifar10 can be done
using the train.py script. Added an example script for training on mnist using
a model function. Densenet is also implemented and can be trained with or
without augmentation. It can only be trained using the script for now.
Resnet is still under development.